+++ src/cmd/services/m3query/config/config.go
@@ -22,7 +22,6 @@ package config
 
 import (
 	"errors"
-	"fmt"
 	"time"
 
 	etcdclient "github.com/m3db/m3/src/cluster/client/etcd"
@@ -38,10 +37,7 @@ import (
 	"github.com/m3db/m3/src/query/storage/m3/consolidators"
 	"github.com/m3db/m3/src/query/storage/m3/storagemetadata"
 	xconfig "github.com/m3db/m3/src/x/config"
-	"github.com/m3db/m3/src/x/config/listenaddress"
-	"github.com/m3db/m3/src/x/cost"
 	"github.com/m3db/m3/src/x/debug/config"
-	xdocs "github.com/m3db/m3/src/x/docs"
 	"github.com/m3db/m3/src/x/instrument"
 	xlog "github.com/m3db/m3/src/x/log"
 	"github.com/m3db/m3/src/x/opentracing"
@@ -62,9 +58,6 @@ const (
 	NoopEtcdStorageType BackendStorageType = "noop-etcd"
 
 	defaultCarbonIngesterListenAddress = "0.0.0.0:7204"
-	errNoIDGenerationScheme            = "error: a recent breaking change means that an ID " +
-		"generation scheme is required in coordinator configuration settings. " +
-		"More information is available here: %s"
 
 	defaultQueryTimeout = 30 * time.Second
 
@@ -117,7 +110,7 @@ type Configuration struct {
 	ClusterManagement *ClusterManagementConfiguration `yaml:"clusterManagement"`
 
 	// ListenAddress is the server listen address.
-	ListenAddress *listenaddress.Configuration `yaml:"listenAddress" validate:"nonzero"`
+	ListenAddress string `yaml:"listenAddress" validate:"nonzero"`
 
 	// Filter is the read/write/complete tags filter configuration.
 	Filter FilterConfiguration `yaml:"filter"`
@@ -167,15 +160,6 @@ type Configuration struct {
 	// StoreMetricsType controls if metrics type is stored or not.
 	StoreMetricsType *bool `yaml:"storeMetricsType"`
 
-	// Cache configurations.
-	//
-	// Deprecated: cache configurations are no longer supported. Remove from file
-	// when we can make breaking changes.
-	// (If/when removed it will make existing configurations with the cache
-	// stanza not able to startup the binary since we parse YAML in strict mode
-	// by default).
-	DeprecatedCache CacheConfiguration `yaml:"cache"`
-
 	// MultiProcess is the multi-process configuration.
 	MultiProcess MultiProcessConfiguration `yaml:"multiProcess"`
 
@@ -215,17 +199,6 @@ type FilterConfiguration struct {
 	Read         Filter `yaml:"read"`
 	Write        Filter `yaml:"write"`
 	CompleteTags Filter `yaml:"completeTags"`
-}
-
-// CacheConfiguration contains the cache configurations.
-type CacheConfiguration struct {
-	// Deprecated: remove from config.
-	DeprecatedQueryConversion *DeprecatedQueryConversionCacheConfiguration `yaml:"queryConversion"`
-}
-
-// DeprecatedQueryConversionCacheConfiguration is deprecated: remove from config.
-type DeprecatedQueryConversionCacheConfiguration struct {
-	Size *int `yaml:"size"`
 }
 
 // ResultOptions are the result options for query.
@@ -324,41 +297,6 @@ func (c PrometheusQueryConfiguration) MaxSamplesPerQueryOrDefault() int {
 type LimitsConfiguration struct {
 	// PerQuery configures limits which apply to each query individually.
 	PerQuery PerQueryLimitsConfiguration `yaml:"perQuery"`
-
-	// Global configures limits which apply across all queries running on this
-	// instance.
-	Global GlobalLimitsConfiguration `yaml:"global"`
-
-	// deprecated: use PerQuery.MaxComputedDatapoints instead.
-	DeprecatedMaxComputedDatapoints int `yaml:"maxComputedDatapoints"`
-}
-
-// MaxComputedDatapoints is a getter providing backwards compatibility between
-// LimitsConfiguration.DeprecatedMaxComputedDatapoints and
-// LimitsConfiguration.PerQuery.PrivateMaxComputedDatapoints. See
-// LimitsConfiguration.PerQuery.PrivateMaxComputedDatapoints for a comment on
-// the semantics.
-func (lc LimitsConfiguration) MaxComputedDatapoints() int {
-	if lc.PerQuery.PrivateMaxComputedDatapoints != 0 {
-		return lc.PerQuery.PrivateMaxComputedDatapoints
-	}
-
-	return lc.DeprecatedMaxComputedDatapoints
-}
-
-// GlobalLimitsConfiguration represents limits on resource usage across a query
-// instance. Zero or negative values imply no limit.
-type GlobalLimitsConfiguration struct {
-	// MaxFetchedDatapoints limits the max number of datapoints allowed to be
-	// used by all queries at any point in time, this is applied at the query
-	// service after the result has been returned by a storage node.
-	MaxFetchedDatapoints int `yaml:"maxFetchedDatapoints"`
-}
-
-// AsLimitManagerOptions converts this configuration to
-// cost.LimitManagerOptions for MaxFetchedDatapoints.
-func (l *GlobalLimitsConfiguration) AsLimitManagerOptions() cost.LimitManagerOptions {
-	return toLimitManagerOptions(l.MaxFetchedDatapoints)
 }
 
 // PerQueryLimitsConfiguration represents limits on resource usage within a
@@ -376,26 +314,6 @@ type PerQueryLimitsConfiguration struct {
 
 	// RequireExhaustive results in an error if the query exceeds any limit.
 	RequireExhaustive *bool `yaml:"requireExhaustive"`
-
-	// MaxFetchedDatapoints limits the max number of datapoints allowed to be
-	// used by a given query, this is applied at the query service after the
-	// result has been returned by a storage node.
-	MaxFetchedDatapoints int `yaml:"maxFetchedDatapoints"`
-
-	// PrivateMaxComputedDatapoints limits the number of datapoints that can be
-	// returned by a query. It's determined purely
-	// from the size of the time range and the step size (end - start / step).
-	//
-	// N.B.: the hacky "Private" prefix is to indicate that callers should use
-	// LimitsConfiguration.MaxComputedDatapoints() instead of accessing
-	// this field directly.
-	PrivateMaxComputedDatapoints int `yaml:"maxComputedDatapoints"`
-}
-
-// AsLimitManagerOptions converts this configuration to
-// cost.LimitManagerOptions for MaxFetchedDatapoints.
-func (l *PerQueryLimitsConfiguration) AsLimitManagerOptions() cost.LimitManagerOptions {
-	return toLimitManagerOptions(l.MaxFetchedDatapoints)
 }
 
 // AsFetchOptionsBuilderLimitsOptions converts this configuration to
@@ -421,13 +339,6 @@ func (l *PerQueryLimitsConfiguration) AsFetchOptionsBuilderLimitsOptions() handl
 		DocsLimit:         int(docsLimit),
 		RequireExhaustive: requireExhaustive,
 	}
-}
-
-func toLimitManagerOptions(limit int) cost.LimitManagerOptions {
-	return cost.NewLimitManagerOptions().SetDefaultLimit(cost.Limit{
-		Threshold: cost.Cost(limit),
-		Enabled:   limit > 0,
-	})
 }
 
 // IngestConfiguration is the configuration for ingestion server.
@@ -489,12 +400,9 @@ type CarbonConfiguration struct {
 
 // CarbonIngesterConfiguration is the configuration struct for carbon ingestion.
 type CarbonIngesterConfiguration struct {
-	// Deprecated: simply use the logger debug level, this has been deprecated
-	// in favor of setting the log level to debug.
-	DeprecatedDebug bool                              `yaml:"debug"`
-	ListenAddress   string                            `yaml:"listenAddress"`
-	MaxConcurrency  int                               `yaml:"maxConcurrency"`
-	Rules           []CarbonIngesterRuleConfiguration `yaml:"rules"`
+	ListenAddress  string                            `yaml:"listenAddress"`
+	MaxConcurrency int                               `yaml:"maxConcurrency"`
+	Rules          []CarbonIngesterRuleConfiguration `yaml:"rules"`
 }
 
 // LookbackDurationOrDefault validates the LookbackDuration
@@ -680,7 +588,7 @@ type TagOptionsConfiguration struct {
 	// If not provided, defaults to `le`.
 	BucketName string `yaml:"bucketName"`
 
-	// Scheme determines the default ID generation scheme. Defaults to TypeLegacy.
+	// Scheme determines the default ID generation scheme. Defaults to TypeQuoted.
 	Scheme models.IDSchemeType `yaml:"idScheme"`
 
 	// Filters are optional tag filters, removing all series with tags
@@ -719,9 +627,8 @@ func TagOptionsFromConfig(cfg TagOptionsConfiguration) (models.TagOptions, error
 	}
 
 	if cfg.Scheme == models.TypeDefault {
-		// If no config has been set, error.
-		docLink := xdocs.Path("how_to/query#migration")
-		return nil, fmt.Errorf(errNoIDGenerationScheme, docLink)
+		// Default to quoted if unspecified.
+		cfg.Scheme = models.TypeQuoted
 	}
 
 	opts = opts.SetIDSchemeType(cfg.Scheme)