+++ src/query/api/v1/handler/namespace/get.go
@@ -43,10 +43,6 @@ import (
 )
 
 var (
-	// DeprecatedM3DBGetURL is the deprecated url for the namespace get handler (with the GET method).
-	// Maintained for backwards compatibility.
-	DeprecatedM3DBGetURL = path.Join(handler.RoutePrefixV1, NamespacePathName)
-
 	// M3DBGetURL is the url for the namespace get handler (with the GET method).
 	M3DBGetURL = path.Join(handler.RoutePrefixV1, M3DBServiceNamespacePathName)